Refactor : 예외 메세지 `출력 구조`를 변경
