+++ src/main/java/christmas/exception/InvalidDateException.java
@@ -1,11 +1,11 @@
 package christmas.exception;
 
-public class InvalidDateException extends ChristmasException {
-    InvalidDateException(String message) {
-        super(message);
+public class InvalidDateException extends IllegalArgumentException {
+    InvalidDateException(ErrorMessage message) {
+        super(message.getMessage());
     }
 
-    public static InvalidDateException of(ErrorMessage error) {
-        return new InvalidDateException(error.getMessage());
+    public static InvalidDateException of() {
+        return new InvalidDateException(ErrorMessage.INVALID_DATE);
     }
 }
