+++ src/main/java/christmas/exception/InvalidOrderException.java
@@ -1,11 +1,11 @@
 package christmas.exception;
 
-public class InvalidOrderException extends ChristmasException {
-    private InvalidOrderException(String message) {
-        super(message);
+public class InvalidOrderException extends IllegalArgumentException {
+    private InvalidOrderException(ErrorMessage message) {
+        super(message.getMessage());
     }
 
-    public static InvalidOrderException of(ErrorMessage error) {
-        return new InvalidOrderException(error.getMessage());
+    public static InvalidOrderException of() {
+        return new InvalidOrderException(ErrorMessage.INVALID_ORDER);
     }
 }
