+++ src/main/java/christmas/util/Validator.java
@@ -1,10 +1,9 @@
 package christmas.util;
 
-import christmas.exception.ErrorMessage;
 import christmas.exception.InvalidOrderException;
 
 public class Validator {
-    public static <T> T validateDuplication(T existedMenu) {
-        throw InvalidOrderException.of(ErrorMessage.ALREADY_EXISTED_VALUE);
+    public static <T> T validateDuplication() {
+        throw InvalidOrderException.of();
     }
 }
